error fixed

+++ src/main/java/com/app/estore/response/CustomerProductDto.java
@@ -5,8 +5,8 @@ import lombok.Data;
 
 @Data
 public class CustomerProductDto {
-    private Integer productId;
-    private Integer vendorId;
+    private Long productId;
+    private Long vendorId;
     private String title;
     private String description;
     private String weight;
